Add files via upload
It seemed like the baseurl="/changdds" messes up the links + header info by adding /changdds before each reference.  I just removed it and made it empty, that way when I did jekyll build it matches what is currently on the website.  I don't know if you changed this on purpose or it was a default add by github... anyways, it seems to work ok for me now.  Let me know if what I did here was wrong in changing baseurl="/changdds" to baseurl=""